Fixed #383: Spring security messages should be english localized in Rest API. Moved batch config and tests into core package so the configurations of batch and batch admin will be separated.
