+++ saos-batch/src/main/java/pl/edu/icm/saos/batch/core/BatchConfiguration.java
@@ -1,4 +1,4 @@
-package pl.edu.icm.saos.batch;
+package pl.edu.icm.saos.batch.core;
 
 import java.util.Set;
 import java.util.TreeSet;
