+++ saos-batch/src/test/java/pl/edu/icm/saos/batch/core/importer/TestCommonCourtsGenerator.java
@@ -1,4 +1,4 @@
-package pl.edu.icm.saos.batch.importer;
+package pl.edu.icm.saos.batch.core.importer;
 
 import java.util.List;
