+++ saos-batch/src/test/java/pl/edu/icm/saos/batch/core/indexer/JudgmentReindexingJobTest.java
@@ -1,4 +1,4 @@
-package pl.edu.icm.saos.batch.indexer;
+package pl.edu.icm.saos.batch.core.indexer;
 
 import static org.junit.Assert.assertEquals;
 
@@ -27,7 +27,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.jdbc.datasource.init.ScriptException;
 
-import pl.edu.icm.saos.batch.BatchTestSupport;
+import pl.edu.icm.saos.batch.core.BatchTestSupport;
 import pl.edu.icm.saos.common.testcommon.category.SlowTest;
 import pl.edu.icm.saos.persistence.common.TestPersistenceObjectFactory;
 import pl.edu.icm.saos.persistence.model.CommonCourtJudgment;
