+++ saos-webapp/src/main/java/pl/edu/icm/saos/webapp/DispatcherServletInitializer.java
@@ -6,7 +6,7 @@ import org.springframework.web.filter.CharacterEncodingFilter;
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
 
 import pl.edu.icm.saos.api.ApiConfiguration;
-import pl.edu.icm.saos.batch.BatchConfiguration;
+import pl.edu.icm.saos.batch.core.BatchConfiguration;
 import pl.edu.icm.saos.common.CommonConfiguration;
 import pl.edu.icm.saos.enrichment.EnrichmentConfiguration;
 import pl.edu.icm.saos.importer.ImportConfiguration;
@@ -19,7 +19,7 @@ public class DispatcherServletInitializer extends AbstractAnnotationConfigDispat
    
     @Override
     protected Class<?>[] getRootConfigClasses() {
-        return new Class<?>[] {  CommonConfiguration.class, EnrichmentConfiguration.class, SecurityConfiguration.class, BatchConfiguration.class, PersistenceConfiguration.class, ImportConfiguration.class,  SearchConfiguration.class, ApiConfiguration.class, GeneralConfiguration.class};
+        return new Class<?>[] {  GeneralConfiguration.class, CommonConfiguration.class, EnrichmentConfiguration.class, SecurityConfiguration.class, BatchConfiguration.class, PersistenceConfiguration.class, ImportConfiguration.class,  SearchConfiguration.class, ApiConfiguration.class};
     }
 
     @Override
