+++ saos-webapp/src/main/resources/message/common.properties
@@ -1,4 +1,4 @@
-saos.fullname = System Analizy Orzeczeń Sądowych
+saos.fullname = Court Judgment Analysis System 
 
 pageNotFound = Page not found
